pyproject.toml
[tool.poetry]
name = "be-teachtech"
version = "0.1.0"
description = ""
authors = ["LE THI HANG <lethihang111001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
tensorflow = "2.17.0"
tensorflow-intel = "2.17.0"
tensorflow-io-gcs-filesystem = "0.31.0"
django = "^5.1.2"
django-jazzmin = "^3.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
mtcnn = "^1.0.0"
keras-facenet = "^0.3.2"
opencv-python = "^4.10.0.84"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
scipy = "^1.14.1"
django-cors-headers = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
